Do some refactoring, fix comments
Vertices and Colors of the triangle are stored in the output buffer.
Additional memory is not used. Data are stored in the output buffer in a certain
order. Vertices are saved in columns, colors in rows. The fields are the same
but there are different functions (methods) to push/pop vertices and
colors into/from the buffer.