Fix error in z calulation for prespective correct texturing
-zEye (after modelvew) should be used for perspective correct texturing
if Frustum is used, but if Ortho is used using -zEye may lead to to division by
zero. So insted of using zEye after modelview Wclip after projection
should be used. For Frustum Wclip will be equal to -zEye. For Ortho Wclip will
be equal to 1.0 so there will not be division by zero for prespective
correct texturing.

[texture integration]