Add field to nm1 tex context for nmgltex functions test
Accessibility of nm1 context from nmpu0 was tested by checking the value
of cntxt_nm1 pointer value (not equal to 0 - nmpu1 is accessible). But
address can be not equal to zero and still be wrong. So to check the
accessibility of nmpu1 for test nmgltex functions an extra field
refValue was added to nm1 texture context.
So if context pointer is not null then the value of refValue can be
checked to be sure that context is acessible.
refValue is only compiled if NMGL_TEX_FUNCS_TEST macro is defined.

Work with hal/x64 branch (commit 0899cd3) and nmpp commit 4e316880

nmglTexFuncs x86 app work with hal/x64 if there is delay between start host,
target0 and target1 applications:
start "" bin\....host.exe
sleep 1
start "" bin\....target0.exe
sleep 1
start "" bin\....target1.exe