Add testperf scenarios for service functions
Add test description to service.h and arrays to corresponding main files
of nmpu0 and nmpu1 test/perf/mai projects.
There are no DDR option in test scenarios (there are arrays in main.cpp files )
because it is slow than internal memory and the performance will be low when
data are in DDR. In order to use DDR arrays add emi_ij array to corresponding
parameter of test scenario. There are four DDR arrays for nmpu0 and two arrays -
for nmpu1.

NOTE:
This commit uses modified vertexPrimitiveRepack.cpp file instead of
.asm. Tests work but rename .cpp to ._cpp and ._asm to .asm in src_proc0/nm to
build the right version of vertexPrimitiveRepack function.