Fix error in textureTriangle with overflowing GL_ADD
* Applying texture function can cause averflow of Cv and Av primary color
values after texturing. In color sum (3.9, gl 1.3) primary and
secondary color are added and clamped to [0;1]. So if primary color is
greater than 1.0 it can be clamped to 1.0 early after texturing
if color sum is not applying in pipeline. Without color sum result color
can be wrong so early test can be useful.