Imlement glEnable/glDisable for texturing
To implement glEnable/glDisable for texturing field textureEnabled was
added to context. Also nmsynchro command was added and functions
glEnable/glDisable was modified. textureEnabled field of context is bit
mask, value 1 of bit means that appropriate texture unit is enabled, 0 -
disabled. textureEanbled >0 if one of all texture units is enabled.
There is one texture unit in nmOpenGL at now. textureEanbled can be
used to check if texturing is enabeld or disabled.