Repack coords, tex coords and colors (without addition in case of odd output)
Repacking works but need to implement insertion of one more column in
case of odd number of output triangles. Function generates even number
of triangles but the last triangle is NOT now a copy of the previous one.

To do:
Copy n-2'th triangle to n-1'th in case of odd number of output triangles

Notes:
* Program stalls when trying to print output arrays before the function
call.
* Sometime the output arrays had only zeroes just like the function
doesn't work at all