Move the assignment of TrianglePointers fields in a separate function
TrianglePointers' fields are the pointers to the data that are stored in
a big buffer. It is necessary to set the pointers to the special places
in that buffer. This code has many lines of code and it will be
duplicated in each test, so move it to a separate function.

Refs #266 @0h14m