textureTriangle nm0: Fix undefined reference to palettes pointers in test
There were undefined reference when build test/nmpu0/textureTriangle
because palettes_p and palettes_widths_p are defined in
nmglvsNm0Init.cpp that is not linked with files in
test/nmpu0/textureTriangle. To fix this errors explicit definitions of
variables were added. Their values are not used in test,
all we need to do is define them.
TODO: remove explicit definitions when the cause of the error will be
removed.