Add FLAT/SMOOTH shade model option to splitTriangles
Color for new vertex of sub triangles were calculated based on all three vertices
of origin triangle that is correct for GL_SMOOTH shade model but not for
GL_FLAT shade model.
As result if shade model is GL_FLAT then wrong colors can be generated.
If shade model is GL_FLAT then color of sub triangles should be
equal to color of third vertex (v2 for v0,v1,v2) of origin triangle.