Create tests description for each test
Correctness tests are based on the parameterized functions. It would be
better to create a description for that parameterized functions
(templates) and then refer to them in the descriptions for the
correctness tests. But it requires refactoring of test description
generator. So the solution is to create multiple descriptions (for each
correctness test). And add a @use tag as a link to the base function.